Setup error when /data/ is not writable

+++ app/I18N.php
@@ -376,7 +376,7 @@ class I18N {
 			}
 		}
 
-		File::mkdir(WT_DATA_DIR . 'cache');
+		$cache_dir_exists = File::mkdir(WT_DATA_DIR . 'cache');
 		$cache_file = WT_DATA_DIR . 'cache/language-' . self::$locale->languageTag() . '-cache.php';
 		if (file_exists($cache_file)) {
 			$filemtime = filemtime($cache_file);
@@ -421,7 +421,9 @@ class I18N {
 				$translation  = new Translation($translation_file);
 				$translations = array_merge($translations, $translation->asArray());
 			}
-			file_put_contents($cache_file, '<' . '?php return ' . var_export($translations, true) . ';');
+			if ($cache_dir_exists) { // During setup, we may not have been able to create it.
+				file_put_contents($cache_file, '<' . '?php return ' . var_export($translations, true) . ';');
+			}
 		} else {
 			$translations = include $cache_file;
 		}
